code/raspberry/requirements.txt
dcd-sdk>=0.0.12
paho-mqtt
python-dotenv
pyserial
requests
pygatt
pexpect
